+++ packages/adapter-static/test/apps/prerendered/svelte.config.js
@@ -3,11 +3,7 @@ import adapter from '../../../index.js';
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
-		adapter: adapter(),
-
-		prerender: {
-			default: true
-		}
+		adapter: adapter()
 	}
 };