+++ packages/kit/src/core/sync/write_types/test/slugs-layout-not-all-pages-have-load/_expected/nested/$types.d.ts
@@ -10,7 +10,7 @@ type OutputDataShape<T> = MaybeWithVoid<
 		Partial<Pick<App.PageData, keyof T & keyof App.PageData>> &
 		Record<string, any>
 >;
-type EnsureParentData<T> = NonNullable<T> extends never ? {} : T;
+type EnsureParentData<T> = T extends null | undefined ? {} : T;
 type LayoutParams = RouteParams & { rest?: string };
 type LayoutParentData = EnsureParentData<import('../$types.js').LayoutData>;