+++ packages/kit/src/runtime/server/page/render.js
@@ -1,4 +1,4 @@
-import devalue from 'devalue';
+import { devalue } from 'devalue';
 import { readable, writable } from 'svelte/store';
 import * as cookie from 'cookie';
 import { hash } from '../../hash.js';
@@ -23,7 +23,7 @@ const updated = {
  *   cookies: import('set-cookie-parser').Cookie[];
  *   options: import('types').SSROptions;
  *   state: import('types').SSRState;
- *   page_config: { hydrate: boolean, router: boolean };
+ *   page_config: { ssr: boolean; csr: boolean };
  *   status: number;
  *   error: HttpError | Error | null;
  *   event: import('types').RequestEvent;
@@ -49,7 +49,7 @@ export async function render_response({
 			throw new Error('Cannot use prerendering if config.kit.csp.mode === "nonce"');
 		}
 
-		if (options.template_contains_nonce) {
+		if (options.app_template_contains_nonce) {
 			throw new Error('Cannot use prerendering if page template contains %sveltekit.nonce%');
 		}
 	}
@@ -75,7 +75,7 @@ export async function render_response({
 		error.stack = options.get_stack(error);
 	}
 
-	if (resolve_opts.ssr) {
+	if (page_config.ssr) {
 		/** @type {Record<string, any>} */
 		const props = {
 			stores: {
@@ -175,6 +175,31 @@ export async function render_response({
 	/** @param {string} path */
 	const prefixed = (path) => (path.startsWith('/') ? path : `${assets}/${path}`);
 
+	const serialized = { data: '', errors: 'null' };
+
+	try {
+		serialized.data = devalue(branch.map(({ server_data }) => server_data));
+	} catch (e) {
+		// If we're here, the data could not be serialized with devalue
+		// TODO if we wanted to get super fancy we could track down the origin of the `load`
+		// function, but it would mean passing more stuff around than we currently do
+		const error = /** @type {any} */ (e);
+		const match = /\[(\d+)\]\.data\.(.+)/.exec(error.path);
+		if (match) throw new Error(`${error.message} (data.${match[2]})`);
+		throw error;
+	}
+
+	if (validation_errors) {
+		try {
+			serialized.errors = devalue(validation_errors);
+		} catch (e) {
+			// If we're here, the data could not be serialized with devalue
+			const error = /** @type {any} */ (e);
+			if (error.path) throw new Error(`${error.message} (errors.${error.path})`);
+			throw error;
+		}
+	}
+
 	// we use an anonymous function instead of an arrow function to support
 	// older browsers (https://github.com/sveltejs/kit/pull/5417)
 	const init_service_worker = `
@@ -218,11 +243,14 @@ export async function render_response({
 		head += `\n\t<link href="${path}" ${attributes.join(' ')}>`;
 	}
 
-	if (page_config.router || page_config.hydrate) {
+	if (page_config.csr) {
 		// Injecting (potentially) legacy script together with the modern script -
 		//  in a similar fashion to the script tags injection of @vitejs/plugin-legacy.
 		// Notice that unlike the script injection on @vitejs/plugin-legacy,
 		//  we don't need to have a constant CSP since kit handles it.
+
+		const detectModernBrowserVarName = '__KIT_is_modern_browser';
+		const startup_script_var_name = '__KIT_startup_script';
 
 		if (modern_polyfills_file) {
 			const path = prefixed(modern_polyfills_file);
@@ -281,35 +309,32 @@ export async function render_response({
 
 		/**
 		 *
-		 * @param {string} prefix
+		 * @param {boolean} detectLegacy If to detect whether we're in legacy, using a global variable.
 		 * @returns
 		 */
-		const getStartupContent = (prefix) =>
-			// prettier-ignore
-			`
-		${prefix}set_public_env(${s(options.public_env)});
-		
-		${prefix}start({
-			target: document.querySelector('[data-sveltekit-hydrate="${target}"]').parentNode,
-			paths: ${s(options.paths)},
-			route: ${!!page_config.router},
-			spa: ${!resolve_opts.ssr},
-			trailing_slash: ${s(options.trailing_slash)},
-			hydrate: ${resolve_opts.ssr && page_config.hydrate ? `{
+		// prettier-ignore
+		const getStartupContent = (detectLegacy) => `
+		start({
+			env: ${s(options.public_env)},
+			hydrate: ${page_config.ssr ? `{
 				status: ${status},
 				error: ${error && serialize_error(error, e => e.stack)},
 				node_ids: [${branch.map(({ node }) => node.index).join(', ')}],
 				params: ${devalue(event.params)},
-				routeId: ${s(event.routeId)}
-			}` : 'null'}
+				routeId: ${s(event.routeId)},
+				data: ${serialized.data},
+				errors: ${serialized.errors}
+			}` : 'null'},
+			paths: ${s(options.paths)},
+			target: document.querySelector('[data-sveltekit-hydrate="${target}"]').parentNode,
+			trailing_slash: ${s(options.trailing_slash)},
+			legacy: ${detectLegacy ? `!window.${detectModernBrowserVarName}` : 'false'}
 		});
 		`;
 
-		const detectModernBrowserVarName = '__KIT_is_modern_browser';
-		const startup_script_var_name = '__KIT_startup_script';
 		if (legacy_entry_file) {
-			const startup_script_js = `window.${startup_script_var_name} = function (m) { ${getStartupContent(
-				'm.'
+			const startup_script_js = `window.${startup_script_var_name} = function (m) { var start = m.start; ${getStartupContent(
+				true
 			)} };`;
 			body += `\n\t\t<script${
 				csp.script_needs_nonce ? ` nonce="${csp.nonce}"` : ''
@@ -348,8 +373,8 @@ export async function render_response({
 			import(${s(prefixed(entry.file))}).then(window.${startup_script_var_name});
 		}
 		` : `
-		import { set_public_env, start } from ${s(prefixed(entry.file))};
-		${getStartupContent('')}`;
+		import { start } from ${s(prefixed(entry.file))};
+		${getStartupContent(false)}`;
 		const attributes = ['type="module"', `data-sveltekit-hydrate="${target}"`];
 
 		csp.add_script(init_app);
@@ -361,7 +386,7 @@ export async function render_response({
 		body += `\n\t\t<script ${attributes.join(' ')}>${init_app}</script>`;
 	}
 
-	if (resolve_opts.ssr && page_config.hydrate) {
+	if (page_config.ssr && page_config.csr) {
 		/** @type {string[]} */
 		const serialized_data = [];
 
@@ -370,15 +395,6 @@ export async function render_response({
 				render_json_payload_script(
 					{ type: 'data', url, body: typeof body === 'string' ? hash(body) : undefined },
 					response
-				)
-			);
-		}
-
-		if (branch.some((node) => node.server_data)) {
-			serialized_data.push(
-				render_json_payload_script(
-					{ type: 'server_data' },
-					branch.map(({ server_data }) => server_data)
 				)
 			);
 		}
@@ -421,7 +437,7 @@ export async function render_response({
 	// TODO flush chunks as early as we can
 	const html =
 		(await resolve_opts.transformPageChunk({
-			html: options.template({ head, body, assets, nonce: /** @type {string} */ (csp.nonce) }),
+			html: options.app_template({ head, body, assets, nonce: /** @type {string} */ (csp.nonce) }),
 			done: true
 		})) || '';