+++ packages/kit/test/apps/amp/svelte.config.js
@@ -1,11 +1,6 @@
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
-		browser: {
-			router: false,
-			hydrate: false
-		},
-
 		inlineStyleThreshold: Infinity
 	}
 };