+++ packages/kit/test/apps/basics/src/routes/errors/error-html/+page.svelte
@@ -1,0 +1,8 @@
+<script>
+	async function fail() {
+		await fetch('./error-html/make-root-fail');
+		location.reload();
+	}
+</script>
+
+<button on:click={fail}>Let's fail catastrophically</button>