+++ packages/kit/test/apps/basics/src/routes/no-csr/+page.svelte
@@ -1,0 +1,7 @@
+<script>
+	/** @type {import('./$types').PageData} */
+	export let data;
+</script>
+
+<h1>look ma no javascript</h1>
+<p>type: {data.type}</p>