+++ packages/kit/types/index.d.ts
@@ -19,6 +19,8 @@ import {
 import { SSRNodeLoader, SSRRoute, ValidatedConfig } from './internal.js';
 import { HttpError, Redirect } from '../src/runtime/control.js';
 
+export { PrerenderOption } from './private.js';
+
 export interface Adapter {
 	name: string;
 	adapt(builder: Builder): MaybePromise<void>;
@@ -99,7 +101,7 @@ export interface Builder {
 	/**
 	 * @param {string} directory Path to the directory containing the files to be compressed
 	 */
-	compress(directory: string): void;
+	compress(directory: string): Promise<void>;
 }
 
 export interface Config {
@@ -121,10 +123,6 @@ export interface KitConfig {
 	adapter?: Adapter;
 	alias?: Record<string, string>;
 	appDir?: string;
-	browser?: {
-		hydrate?: boolean;
-		router?: boolean;
-	};
 	csp?: {
 		mode?: 'hash' | 'nonce' | 'auto';
 		directives?: CspDirectives;
@@ -142,7 +140,8 @@ export interface KitConfig {
 		params?: string;
 		routes?: string;
 		serviceWorker?: string;
-		template?: string;
+		appTemplate?: string;
+		errorTemplate?: string;
 	};
 	inlineStyleThreshold?: number;
 	methodOverride?: {
@@ -260,7 +259,6 @@ export interface RequestHandler<
 }
 
 export interface ResolveOptions {
-	ssr?: boolean;
 	transformPageChunk?: (input: { html: string; done: boolean }) => MaybePromise<string | undefined>;
 }