+++ packages/migrate/migrations/routes/utils.js
@@ -494,3 +494,38 @@ export function rewrite_type(node, code, old_type, new_type) {
 		code.overwrite(start, start + old_type.length, new_type);
 	}
 }
+
+/**
+ * Does the HTTP verbs uppercase migration if it didn't happen yet. If a string
+ * is returned, the migration was done or wasn't needed. If undefined is returned,
+ * the migration is needed but couldn't be done.
+ *
+ * @param {string[]} methods
+ * @param {NonNullable<ReturnType<typeof parse>>} file
+ */
+export function uppercase_migration(methods, file) {
+	const old_methods = ['get', 'post', 'put', 'patch', 'del'].filter((name) =>
+		file.exports.map.has(name)
+	);
+
+	if (old_methods.length && !methods.length) {
+		for (const statement of file.ast.statements) {
+			for (const method of old_methods) {
+				const fn = get_function_node(
+					statement,
+					/** @type{string} */ (file.exports.map.get(method))
+				);
+				if (!fn?.name) {
+					return;
+				}
+				file.code.overwrite(
+					fn.name.getStart(),
+					fn.name.getEnd(),
+					method === 'del' ? 'DELETE' : method.toUpperCase()
+				);
+			}
+		}
+	}
+
+	return file.code.toString();
+}