+++ packages/adapter-auto/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sveltejs/adapter-auto
 
+## 1.0.0-next.70
+
+### Patch Changes
+
+- Updated dependencies [[`4b48d1bd`](https://github.com/sveltejs/kit/commit/4b48d1bd5005dc8e059b17fa42f0d08674aad30c)]:
+  - @sveltejs/adapter-vercel@1.0.0-next.72
+  - @sveltejs/adapter-netlify@1.0.0-next.75
+
+## 1.0.0-next.69
+
+### Patch Changes
+
+- Updated dependencies [[`c530d337`](https://github.com/sveltejs/kit/commit/c530d33793a228fac684c71ed7926e6217101a90)]:
+  - @sveltejs/adapter-netlify@1.0.0-next.75
+  - @sveltejs/adapter-vercel@1.0.0-next.71
+
+## 1.0.0-next.68
+
+### Patch Changes
+
+- Updated dependencies [[`331e0a80`](https://github.com/sveltejs/kit/commit/331e0a80b5e63f75c8226ddc93d0b0beefd40de3)]:
+  - @sveltejs/adapter-vercel@1.0.0-next.70
+  - @sveltejs/adapter-netlify@1.0.0-next.74
+
 ## 1.0.0-next.67
 
 ### Patch Changes