+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sveltejs/adapter-vercel
 
+## 1.0.0-next.72
+
+### Patch Changes
+
+- Fix `global is not defined` runtime error building with `edge: true` option ([#6425](https://github.com/sveltejs/kit/pull/6425))
+
+## 1.0.0-next.71
+
+### Patch Changes
+
+- Use devalue to serialize server-only `load` return values ([#6318](https://github.com/sveltejs/kit/pull/6318))
+
+## 1.0.0-next.70
+
+### Patch Changes
+
+- Don't use top-level-await, as it is not supported right now ([#6360](https://github.com/sveltejs/kit/pull/6360))
+
 ## 1.0.0-next.69
 
 ### Patch Changes